9C-1084: Stores packages into collection table

+++ modules/commons/src/main/scala/cards/nine/domain/pagination/Page.scala
@@ -1,3 +1,3 @@
 package cards.nine.domain.pagination
 
-case class Page(pageNumber: Int, pageSize: Int)
+case class Page(pageNumber: Long, pageSize: Long)
